requirements.txt
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.5
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
python-jose==3.3.0
openai>=1.10.0,<2.0.0
sqlalchemy==2.0.23
pydantic==2.5.2
